requirements.txt
mcp
notion-client
python-dotenv
starlette<0.47.0


